backend/requirements.txt
Django==3.2
flake8==5.0.4
djangorestframework==3.12.4
mccabe==0.7.0
pyflakes==2.5.0
pytz==2023.3
snowballstemmer==2.2.0
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
Pillow==8.3.1
djoser==2.1.0
drf-extra-fields==3.4.0
django-filter==22.1
